Factor out dirs.ts.
